+++ kifi-backend/modules/common/conf/prod/prod.conf
@@ -154,9 +154,9 @@ ws {
   useragent = "Play42"
 }
 
-# ~~~~~~~~~~~~~~~~~~~~~~
-#      Scraper config
-# ~~~~~~~~~~~~~~~~~~~~~~
+# ~~~~~~~~~~~~~~~~~~~~~~~~
+#  Scraper Schedule config
+# ~~~~~~~~~~~~~~~~~~~~~~~~
 
 scraper {
 
@@ -167,31 +167,11 @@ scraper {
     decrement = 2                   # hours
   }
 
-  queue {
-    terminateThreshold = 120000
-    sizeThreshold = 100
-    terminatorFreq = 5
-  }
-
-  http {
-    fetcherEnforcerFreq = 5
-    fetcherQSizeThreshold = 100
-  }
-
   initialBackoff = 3                 # hours
   maxBackoff = 1024                  # hours
   maxRandomDelay = 600               # seconds
-  changeThreshold = 0.5
-  pullMultiplier = 8
-  pullFrequency = 5                  # seconds
   scrapePendingFrequency = 30        # seconds
-  queued = true
-  async = false
   actorTimeout = 20000
-  syncAwaitTimeout = 20000
-  serviceCallTimeout = 20000
-  batchSize = 10
-  batchMax = 50
   pendingOverdueThreshold = 20       # minutes
   checkOverdueCountFrequency = 20    # minutes
   overdueCountThreshold = 1000
