+++ kifi-backend/modules/common/conf/prod/scraper.conf
@@ -27,13 +27,6 @@ cdn {
 
 scraper {
 
-  interval {
-    min = 24                        # hours
-    max = 1024                      # hours
-    increment = 6                   # hours
-    decrement = 2                   # hours
-  }
-
   queue {
     terminateThreshold = 120000
     sizeThreshold = 100
@@ -45,23 +38,15 @@ scraper {
     fetcherQSizeThreshold = 100
   }
 
-  initialBackoff = 3                 # hours
-  maxBackoff = 1024                  # hours
-  maxRandomDelay = 600               # seconds
   changeThreshold = 0.5
   pullMultiplier = 8
   pullFrequency = 5                  # seconds
-  scrapePendingFrequency = 30        # seconds
   queued = true
   async = false
-  actorTimeout = 20000
   syncAwaitTimeout = 20000
   serviceCallTimeout = 20000
   batchSize = 10
   batchMax = 50
-  pendingOverdueThreshold = 20       # minutes
-  checkOverdueCountFrequency = 20    # minutes
-  overdueCountThreshold = 1000
 }
 
 # ~~~~~~~~~~~~~~~~~~~~~~
