+++ kifi-backend/modules/common/conf/prod/shoebox.conf
@@ -69,18 +69,3 @@ urban-airship {
 airbrake {
   key = "701215c3c9029a56df52250ce2c5750f"
 }
-
-# ~~~~~~~~~~~~~~~~~~
-# scraper schedule
-# ~~~~~~~~~~~~~~~~~~
-
-scraper {
-  actorTimeout = 20000
-  scrapePendingFrequency = 30        # seconds
-  pendingOverdueThreshold = 20       # minutes
-  checkOverdueCountFrequency = 20    # minutes
-  overdueCountThreshold = 1000
-}
-
-
-
