+++ kifi-backend/modules/scraper/app/com/keepit/scraper/ScrapeProcessorModule.scala
@@ -18,16 +18,17 @@ case class ProdScraperProcessorModule() extends ScrapeProcessorModule {
     bind[SyncShoeboxDbCallbacks].to[ShoeboxDbCallbackHelper].in[AppScoped]
     bind[PullerPlugin].to[PullerPluginImpl].in[AppScoped]
     install(ProdScraperConfigModule())
+    install(ProdScrapeSchedulerConfigModule())
   }
 
   @Singleton
   @Provides
-  def httpFetcher(airbrake:AirbrakeNotifier, schedulingProperties:SchedulingProperties, scraperConfig: ScraperConfig): HttpFetcher = {
+  def httpFetcher(airbrake:AirbrakeNotifier, schedulingProperties:SchedulingProperties, scraperConfig: ScraperConfig, scheduler: ScraperSchedulerConfig): HttpFetcher = {
     new HttpFetcherImpl(
       airbrake,
       userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
-      connectionTimeout = scraperConfig.scrapePendingFrequency * 1000, // todo: revisit
-      soTimeOut = scraperConfig.scrapePendingFrequency * 1000,
+      connectionTimeout = scheduler.scrapePendingFrequency * 1000, // todo: revisit
+      soTimeOut = scheduler.scrapePendingFrequency * 1000,
       trustBlindly = true,
       schedulingProperties,
       scraperConfig.httpConfig
