+++ kifi-backend/modules/scraper/app/com/keepit/scraper/ScraperConfig.scala
@@ -13,30 +13,11 @@ case class ScraperHttpConfig(
   httpFetcherQSizeThreshold: Int
 )
 
-case class ScraperIntervalConfig(
-  minInterval: Double, //hours
-  maxInterval: Double, //hours
-  intervalIncrement: Double, //hours
-  intervalDecrement: Double //hours
-)
-
-case class ScraperSchedulerConfig(
-  actorTimeout: Int,
-  scrapePendingFrequency: Int,          // seconds
-  checkOverdueCountFrequency: Int,      // minutes
-  pendingOverdueThreshold: Int,         // minutes
-  overdueCountThreshold: Int
-)
 
 case class ScraperConfig(
-  intervalConfig: ScraperIntervalConfig,
-  initialBackoff: Double, //hours
-  maxBackoff: Double, //hours
-  maxRandomDelay: Int, // seconds
   changeThreshold: Double,
   pullMultiplier:Int,
   pullFrequency: Int, // seconds
-  scrapePendingFrequency: Int, // seconds
   queued: Boolean,
   async: Boolean,
   syncAwaitTimeout: Int,
@@ -45,12 +26,7 @@ case class ScraperConfig(
   batchMax: Int,
   httpConfig: ScraperHttpConfig,
   queueConfig: ScraperQueueConfig
-) {
-
-  private[this] val rnd = new Random
-
-  def randomDelay(): Int = rnd.nextInt(maxRandomDelay) // seconds
-}
+)
 
 object ScraperConfig {
   val BATCH_SIZE = 100
