+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/internal/EmailNotificationsController.scala
@@ -3,7 +3,6 @@ package com.keepit.controllers.internal
 import com.keepit.common.time.internalTime.DateTimeJsonLongFormat
 import com.google.inject.Inject
 import com.keepit.common.time.Clock
-import com.keepit.scraper.ScraperConfig
 import com.keepit.common.service.FortyTwoServices
 import com.keepit.common.controller.ShoeboxServiceController
 import com.keepit.common.logging.Logging
