+++ kifi-backend/modules/shoebox/app/com/keepit/scraper/ScrapeSchedulerModule.scala
@@ -24,35 +24,3 @@ case class DevScrapeSchedulerModule() extends ScrapeSchedulerModule {
   }
 }
 
-trait ScrapeSchedulerConfigModule extends ScalaModule {
-  protected def conf: Configuration
-}
-
-case class ProdScrapeSchedulerConfigModule() extends ScrapeSchedulerConfigModule {
-  def configure() {}
-
-  override protected def conf: Configuration = Play.current.configuration
-
-  @Singleton
-  @Provides
-  def schedulerConfig(): ScraperSchedulerConfig = {
-    ScraperSchedulerConfig(
-      actorTimeout = conf.getInt("scraper.actorTimeout").get,
-      scrapePendingFrequency = conf.getInt("scraper.scrapePendingFrequency").get,
-      checkOverdueCountFrequency = conf.getInt("scraper.checkOverdueCountFrequency").get,
-      pendingOverdueThreshold = conf.getInt("scraper.pendingOverdueThreshold").get,
-      overdueCountThreshold = conf.getInt("scraper.overdueCountThreshold").get
-    )
-  }
-}
-
-case class TestScrapeSchedulerConfigModule() extends ScrapeSchedulerConfigModule {
-  def configure() {}
-
-  override protected def conf: Configuration = new Configuration(Configuration.empty.underlying) {
-    override def getInt(key: String): Option[Int] = Some(0)
-    override def getDouble(key: String): Option[Double] = Some(0)
-    override def getBoolean(key: String): Option[Boolean] = Some(true)
-  }
-}
-
